+++ common/src/main/java/com/github/scotsguy/nowplaying/gui/toast/NowPlayingToast.java
@@ -37,21 +37,38 @@ import java.util.List;
 import static com.github.scotsguy.nowplaying.util.Localization.localized;
 
 public class NowPlayingToast implements Toast {
-    private static final ResourceLocation BACKGROUND_SPRITE = ResourceLocation.withDefaultNamespace("toast/recipe");
-
+    private static final ResourceLocation BACKGROUND_SPRITE_LIGHT = ResourceLocation.withDefaultNamespace("toast/recipe");
+    private static final int NOW_PLAYING_COLOR_LIGHT = 0xFF800080;
+    private static final int TITLE_COLOR_LIGHT = 0xFF000000;
+    private static final ResourceLocation BACKGROUND_SPRITE_DARK = ResourceLocation.withDefaultNamespace("toast/advancement");
+    private static final int NOW_PLAYING_COLOR_DARK = 0xFF993299;
+    private static final int TITLE_COLOR_DARK = 0xFFD1D1D1;
+    
     private final Component description;
     private final ResourceLocation discSprite;
     private final long displayTime;
     private final float scale;
+    private final ResourceLocation sprite;
+    private final int nowPlayingColor;
+    private final int titleColor;
 
     private static final int TEXT_LEFT_MARGIN = 30;
     private static final int TEXT_RIGHT_MARGIN = 7;
 
-    public NowPlayingToast(Component description, ResourceLocation discSprite, long displayTime, float scale) {
+    public NowPlayingToast(Component description, ResourceLocation discSprite, long displayTime, float scale, boolean darkMode) {
         this.description = description;
         this.discSprite = discSprite;
         this.displayTime = displayTime;
         this.scale = scale;
+        if (darkMode) {
+            sprite = BACKGROUND_SPRITE_DARK;
+            nowPlayingColor = NOW_PLAYING_COLOR_DARK;
+            titleColor = TITLE_COLOR_DARK;
+        } else {
+            sprite = BACKGROUND_SPRITE_LIGHT;
+            nowPlayingColor = NOW_PLAYING_COLOR_LIGHT;
+            titleColor = TITLE_COLOR_LIGHT;
+        }
     }
 
     @Override
@@ -69,7 +86,7 @@ public class NowPlayingToast implements Toast {
 
         if (width == 160 && textLines.size() <= 1) {
             // Text fits, draw the whole toast from the texture
-            graphics.blitSprite(BACKGROUND_SPRITE, 0, 0, width, height);
+            graphics.blitSprite(sprite, 0, 0, width, height);
         } else {
             // Stretch toast by drawing the sprite multiple times
             height = height + Math.max(0, textLines.size() - (Config.options().simpleToast ? 2 : 1)) * 12;
@@ -90,15 +107,15 @@ public class NowPlayingToast implements Toast {
         if (Config.options().simpleToast) {
             // Draw song title only
             for (int i = 0; i < textLines.size(); ++i) {
-                graphics.drawString(mc.font, textLines.get(i), TEXT_LEFT_MARGIN, (textLines.size() == 1 ? 12 : 6 + i * 12), -16777216, false);
+                graphics.drawString(mc.font, textLines.get(i), TEXT_LEFT_MARGIN, (textLines.size() == 1 ? 12 : 6 + i * 12), titleColor, false);
             }
         } else {
             // Draw "Now Playing"
-            graphics.drawString(mc.font, localized("message", "nowPlaying"), TEXT_LEFT_MARGIN, 7, -11534256, false);
+            graphics.drawString(mc.font, localized("message", "nowPlaying"), TEXT_LEFT_MARGIN, 7, nowPlayingColor, false);
 
             // Draw song title
             for (int i = 0; i < textLines.size(); ++i) {
-                graphics.drawString(mc.font, textLines.get(i), TEXT_LEFT_MARGIN, (18 + i * 12), -16777216, false);
+                graphics.drawString(mc.font, textLines.get(i), TEXT_LEFT_MARGIN, (18 + i * 12), titleColor, false);
             }
         }
 
@@ -113,12 +130,12 @@ public class NowPlayingToast implements Toast {
         int uWidth = vOffset == 0 ? 20 : 5;
         int n = Math.min(60, i - uWidth);
 
-        graphics.blitSprite(BACKGROUND_SPRITE, 160, 32, 0, vOffset, 0, y, uWidth, vHeight);
+        graphics.blitSprite(sprite, 160, 32, 0, vOffset, 0, y, uWidth, vHeight);
 
         for (int o = uWidth; o < i - n; o += 64) {
-            graphics.blitSprite(BACKGROUND_SPRITE, 160, 32, 32, vOffset, o, y, Math.min(64, i - o - n), vHeight);
+            graphics.blitSprite(sprite, 160, 32, 32, vOffset, o, y, Math.min(64, i - o - n), vHeight);
         }
 
-        graphics.blitSprite(BACKGROUND_SPRITE, 160, 32, 160 - n, vOffset, i - n, y, n, vHeight);
+        graphics.blitSprite(sprite, 160, 32, 160 - n, vOffset, i - n, y, n, vHeight);
     }
 }