+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.5.13-beta.1
 
 - Added support for defining custom disc sprite locations
+- Added option for dark mode toast
 
 ## 1.5.12